removed some things from QH stuff. Changed test sets

+++ src/rendering.h
@@ -121,7 +121,7 @@ struct Mesh
     GLuint VAO;
     GLuint VBO;
     
-    int vertexCount;
+    size_t vertexCount;
 };
 
 struct DebugContext
